remove useless file
